+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Lightstreamer Changelog
 
-### 0.9 — Unreleased
+### 0.9 — August 4, 2016
 
 - Replaced the `Lightstreamer::Session#error` attribute with a new `Lightstreamer::Session#on_error` callback to enable
   more efficient error handling and notification